fix : move some tests to test_cli.py
