+++ tests/test_cli.py
@@ -4,8 +4,32 @@ from unittest.mock import patch
 import pytest
 from mytext import Provider
 from mytext.cli import main
+from mytext.params import MY_TEXT_VERSION, MY_TEXT_OVERVIEW, MY_TEXT_REPO
 
 TEST_CASE_NAME = "CLI tests"
+
+
+def test_main_version(capsys):
+    with patch("sys.argv", ["mytext", "--version"]):
+        main()
+    out, _ = capsys.readouterr()
+    assert MY_TEXT_VERSION in out
+
+
+def test_main_info(capsys):
+    with patch("sys.argv", ["mytext", "--info"]):
+        main()
+    out, _ = capsys.readouterr()
+    assert MY_TEXT_OVERVIEW in out
+    assert MY_TEXT_REPO in out
+
+
+def test_main_no_text(capsys):
+    with patch("sys.argv", ["mytext"]):
+        with pytest.raises(SystemExit):
+            main()
+    _, err = capsys.readouterr()
+    assert "--text is required" in err
 
 
 @patch("mytext.cli._load_auth_from_env")
